@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Reflection;
+using System.Runtime.CompilerServices;
 using Serilog;
 
 namespace SWTools.Core {
@@ -8,32 +9,35 @@ namespace SWTools.Core {
     /// </summary>
     public static partial class Helper {
         private static readonly AssemblyName _assemblyName = Assembly.GetExecutingAssembly().GetName();
-        private const string _versionUrl = "https://raw.githubusercontent.com/King-zzk/king-zzk.github.io/refs/heads/main/version.txt";
 
         public static readonly Version EmptyVersion = new();
 
         // 获取项目版本信息
-        public static Version GetVersion() {
+        public static Version Version {
+            get {
             if (_assemblyName.Version == null) {
                 return EmptyVersion;
             }
             return _assemblyName.Version;
+            }
         }
 
         // 获取项目版本信息字符串
-        public static string GetVersionStr() {
-            return GetVersion().ToString(3);
+        public static string VersionStr {
+            get {
+                return Version.ToString(3);
+            }
         }
 
         // 获取最新版本
-        public static Version RequestLatestVersion() {
+        public async static Task<Version> RequestLatestVersion() {
             string str;
-            foreach (var proxy in _githubProxies) {
-                str = Helper.MakeHttpGet(proxy + _versionUrl);
+            foreach (var proxy in Constants.UrlGithubProxy) {
+                str = await Helper.MakeHttpGet(proxy + Constants.UrlVersion);
                 if (str != string.Empty) {
                     if (Version.TryParse(str, out Version? version)) {
-                        Log.Logger.Debug("Received version info from \"{Url}\" (version {Version})",
-                            proxy + _versionUrl, version);
+                        LogManager.Log.Debug("Received version info from \"{Url}\" (version {Version})",
+                            proxy + Constants.UrlVersion, version);
                         return version;
                     } else {
                         continue;
