@@ -1,50 +1,25 @@
 ﻿using System;
-using System.Text.Encodings.Web;
-using System.Text.Json.Serialization;
-using System.Text.Json;
-using System.Text.Unicode;
 using System.IO.Compression;
-using Serilog;
 using System.Diagnostics;
 using System.Text;
+using Serilog;
 
 namespace SWTools.Core {
     /// <summary>
     /// 辅助方法
     /// </summary>
     public static partial class Helper {
-        private const string _steamcmdUrl = "https://steamcdn-a.akamaihd.net/client/installer/steamcmd.zip";
-
-        // Json 选项
-        public static readonly JsonSerializerOptions _jsonOptions = new() {
-            // 序列化选项
-            WriteIndented = true,
-            Encoder = JavaScriptEncoder.Create(UnicodeRanges.BasicLatin, UnicodeRanges.CjkUnifiedIdeographs),
-            // 反序列化选项
-            PropertyNameCaseInsensitive = true,
-            UnmappedMemberHandling = JsonUnmappedMemberHandling.Skip,
-            UnknownTypeHandling = JsonUnknownTypeHandling.JsonElement,
-        };
-
-        // Gitub 代理
-        public static readonly string[] _githubProxies = [
-            "", // 这个空的不能删
-            "https://gh.llkk.cc/",
-            "https://gitproxy.click/",
-            "https://ghproxy.net/"
-            // 3 个够了
-        ];
 
         // 安装 Steamcmd
-        public static bool SetupSteamcmd() {
+        public static async Task<bool> SetupSteamcmd() {
             // 安装目录存在
             if (Directory.Exists(ConfigManager.Config.SteamcmdPath)) {
-                Log.Logger.Warning("Directory \"{StramcmdPath}\" already exists, skipping download",
+                LogManager.Log.Warning("Directory \"{StramcmdPath}\" already exists, skipping download",
                     ConfigManager.Config.SteamcmdPath);
             } else { // 下载
                 Directory.CreateDirectory(ConfigManager.Config.SteamcmdPath);
-                if (!DownloadFile(_steamcmdUrl, ConfigManager.Config.SteamcmdPath + "temp.zip")) {
-                    Log.Logger.Error("Failed to download");
+                if (!await DownloadFile(Constants.UrlSteamcmd, ConfigManager.Config.SteamcmdPath + "temp.zip")) {
+                    LogManager.Log.Error("Failed to download");
                     return false;
                 }
                 try {
@@ -59,7 +34,7 @@ namespace SWTools.Core {
                     Log.Information("Installed steamcmd at \"{SteamcmdPath}\"", ConfigManager.Config.SteamcmdPath);
                 }
                 catch (Exception ex) {
-                    Log.Logger.Error("Exception occured when installing Steamcmd:\n{Exception}", ex);
+                    LogManager.Log.Error("Exception occured when installing Steamcmd:\n{Exception}", ex);
                     return false;
                 }
             }
@@ -69,7 +44,7 @@ namespace SWTools.Core {
                 using Process? process = Process.Start(startInfo) ?? throw new Exception("process is null");
                 process.OutputDataReceived += (s, e) => {
                     if (!string.IsNullOrEmpty(e.Data)) {
-                        Log.Logger.Debug("Steamcmd: {Message}", e.Data);
+                        LogManager.Log.Debug("Steamcmd: {Message}", e.Data);
                     }
                 };
                 process.Start();
@@ -77,10 +52,10 @@ namespace SWTools.Core {
                     process.BeginOutputReadLine();
                 }
                 process.WaitForExit();
-                Log.Logger.Information("Completed update of steamcmd");
+                LogManager.Log.Information("Completed update of steamcmd");
             }
             catch (Exception ex) {
-                Log.Logger.Error("Exception occured when launching Steamcmd:\n{Exception}", ex);
+                LogManager.Log.Error("Exception occured when launching Steamcmd:\n{Exception}", ex);
                 return false;
             }
             return true;
@@ -101,6 +76,22 @@ namespace SWTools.Core {
                 StandardErrorEncoding = Encoding.UTF8,
                 StandardInputEncoding = Encoding.UTF8,
             };
+        }
+
+        // 忽略后缀名地查找文件 (返回文件名)
+        public static string? FindFileIgnoreExt(string path, string fileNameWithoutExt) {
+            if (!Directory.Exists(path)) return null;
+            string[] allFiles = Directory.GetFiles(path);
+            foreach (string file in allFiles) {
+                string name = Path.GetFileName(file);
+                string nameNoExt = Path.GetFileNameWithoutExtension(file);
+                if (nameNoExt.Equals(
+                    fileNameWithoutExt,
+                    StringComparison.OrdinalIgnoreCase)) {
+                    return name;
+                }
+            }
+            return null;
         }
     }
 }
