@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
@@ -10,25 +8,24 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+using System.IO;
 
 namespace SWTools.WPF {
     /// <summary>
-    /// MoreWindow.xaml 的交互逻辑
+    /// LogWindow.xaml 的交互逻辑
     /// </summary>
-    public partial class MoreWindow : Window {
-        public MoreWindow() {
+    public partial class LogWindow : Window {
+        public LogWindow() {
             InitializeComponent();
-            VersionText.Text = $"这是 Steam Workshop Tools v{Core.Helper.VersionStr}。";
         }
 
         private void BtnOk_Click(object sender, RoutedEventArgs e) {
             Close();
         }
 
-        private void BtnGithub_Click(object sender, RoutedEventArgs e) {
+        private void BtnOpenFolder_Click(object sender, RoutedEventArgs e) {
             System.Diagnostics.Process.Start("explorer.exe",
-                "https://github.com/King-zzk/Steam-Workshops-Tools-SWTools");
+                Path.GetFullPath(Core.Constants.LogDirName));
         }
     }
 }
